added round_offset to resume training

+++ fed_exp_ptb.py
@@ -22,6 +22,7 @@ from flwr.server import Server
 from flwr.server.client_manager import SimpleClientManager
 
 from utils.model import gen_weighted_avergage_fn, get_params
+from utils.general import find_round_offset
 
 from convert_vital_signs_to_arrow import create_ptb_dataset
 import json
@@ -52,10 +53,12 @@ num_rounds = 10
 
 log_path = args.log_path #"logs/fed_avg_hetro2"
 
+round_offset = find_round_offset(log_path)
+
 if not os.path.exists(log_path):
     os.mkdir(log_path)
 
-if args.strategy == "scaffold" and args.cv_dir != "":
+if round_offset == 0 and args.strategy == "scaffold" and args.cv_dir != "":
     if os.path.exists(args.cv_dir):
         for file in os.listdir(args.cv_dir):
             os.remove(os.path.join(args.cv_dir, file))
@@ -63,7 +66,9 @@ if args.strategy == "scaffold" and args.cv_dir != "":
     os.makedirs(args.cv_dir)
 
 event_file = os.path.join(log_path, "events.txt")
-open(event_file, "w").close()
+
+if round_offset == 0:
+    open(event_file, "w").close()
 
 
 with open(partition_path, "r") as f:
@@ -128,7 +133,8 @@ strategy = strategy_class(
     min_evaluate_clients=len(client_ds),
     min_available_clients=len(client_ds),
     log_path=log_path,
-    event_file=event_file
+    event_file=event_file,
+    round_offset=round_offset
 )
 
 if args.strategy == "scaffold":
